Add files via upload
Main File: ios-data-parser.py.  This can be used to import the biome data from iOS 16+ into a SQLite database along with data from KnowledgeC for user activity querying like was capable with the inFocus data prior to iOS 16.  It also will take the Cache.db and import device locations then generate a KML isolating items with a limited horizontal precision and time period.  It will include the precision in the title, unlike what's in Cellebrite's PA.  You can also import a CSV so if you export one with deleted but recovered record entries, they can be added to the KML the same way.  An optional extra tool, kml-refine.py, will allow you to take an exported KML file and eliminate many of the titles, placing them in the description AND calculate speeds based on the point to point and high and low precisions.  The speeds are in the descriptions for the lines placed between points.  This is good for showing a pattern of behavior of speed without calculating each individual point by hand.